+++ Runtime/DB/Item DB/ArmorItem.cs
@@ -22,7 +22,7 @@ namespace OcDialogue
         public int MaxDurability => maxDurability;
         public float CurrentDurability { get; set; }
         public float Weight => weight;
-        public float Equipped { get; set; }
+        public bool Equipped { get; set; }
 
         public int maxDurability = 100;
         [InlineButton("CalcDefenseFromWeight", "무게로 방어력 계산")][InlineButton("MoreDefense", "▶")][InlineButton("LessDefense", "◀")]
@@ -86,8 +86,23 @@ namespace OcDialogue
 
             copy.stability = stability;
             copy.avatar = avatar;
-            
         }
+        
+        
+        public override bool IsNowUsable()
+        {
+            //TODO:장비 못 하는 경우가 있을지 모르겠다.
+            return true;
+        }
+        public override void Use()
+        {
+            if(!IsNowUsable()) return;
+            //TODO : 장착에 관한 것.
+        }
+        
+        
+        
+        
 #if UNITY_EDITOR
         public override void SetSubTypeFromString(string subtypeName)
         {